Fix some warnings of possible loss of data in casting.
Implemented Load Method of Hydrographic Network in order to test.
add #define WIN32_LEAN_AND_MEAN in order to compile in x64 Platform.

+++ Projects/Project1/src/loader.h
@@ -38,8 +38,8 @@ Loadable* Loader<Loadable>::Load()
     if (!File::Load(_fileName.c_str(), buffer, size))
         return NULL;
 
-    ByteBuffer bb(size);
-    bb.WriteBuffer(buffer, size);
+    ByteBuffer bb(static_cast<uint32>(size));
+    bb.WriteBuffer(buffer, static_cast<uint32>(size));
 
     delete[] buffer;
